+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * In `Nested()`, we no longer use `Railway::End::Success` and `Railway::End::Failure` as static outputs but
   simply use `Railway`'s default termini objects.
+* Use `dsl`'s new `Inject()` API instead of a overriding `:inject` in `Model()` and `Policy()`.
 
 # 2.1.9